# ASP.Net/Commerce_en_ligne/CommerceAPI_NetCore/CommerceAPI_NetCore/Controllers/ProductsController.cs
using CommerceAPI_NetCore.Models;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace CommerceAPI_NetCore.Controllers {

    [Route("api/[controller]")]
    [ApiController]
    public class ProductsController : ControllerBase {
        private readonly EcommerceSimplifieContext _context;

        public ProductsController(EcommerceSimplifieContext context) {
            _context = context;
        }

        // GET: api/Products
        [HttpGet]
        public async Task<ActionResult<IEnumerable<Product>>> GetProducts() {
            return await _context.Products.ToListAsync();
        }

        // GET: api/Products/5
        [HttpGet("{id}")]
        public async Task<ActionResult<Product>> GetProduct(int id) {
            var product = await _context.Products.FindAsync(id);

            if (product == null) {
                return NotFound();
            }

            return product;
        }

        // PUT: api/Products/5
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPut("{id}")]
        public async Task<IActionResult> PutProduct(int id, Product product) {
            if (id != product.Id) {
                return BadRequest();
            }

            _context.Entry(product).State = EntityState.Modified;

            try {
                await _context.SaveChangesAsync();
            } catch (DbUpdateConcurrencyException) {
                if (!ProductExists(id)) {
                    return NotFound();
                } else {
                    throw;
                }
            }

            return NoContent();
        }

        // POST: api/Products
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPost]
        public async Task<ActionResult<Product>> PostProduct(Product product) {
            _context.Products.Add(product);
            await _context.SaveChangesAsync();

            return CreatedAtAction("GetProduct", new {
                id = product.Id
            }, product);
        }

        // DELETE: api/Products/5
        [HttpDelete("{id}")]
        public async Task<IActionResult> DeleteProduct(int id) {
            var product = await _context.Products.FindAsync(id);
            if (product == null) {
                return NotFound();
            }

            _context.Products.Remove(product);
            await _context.SaveChangesAsync();

            return NoContent();
        }

        private bool ProductExists(int id) {
            return _context.Products.Any(e => e.Id == id);
        }

        // GET: api/Products/5/picture
        [HttpGet("{id}/picture")]
        public async Task<ActionResult<Picture>> GetPictureById(int id) {
            var picture = await (from pic in _context.Pictures
                                 join map in _context.ProductPictureMappings on pic.Id equals map.PictureId
                                 select map)
                                 .Where(x => x.ProductId == id && x.DisplayOrder == 1)
                                 .Select(s => s.Picture).FirstOrDefaultAsync();

            if (picture == null) {
                return NotFound();
            }

            return picture;
        }
    }
}