# ASP.Net/Commerce_en_ligne/WebApp/WebApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CommerceAPI_NetCore",
    "version": "v1"
  },
  "paths": {
    "/api/Categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "GetCategories",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "operationId": "PostCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      }
    },
    "/api/Categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "GetCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "operationId": "PutCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "operationId": "DeleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Categories/{id}/children": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "GetChildrenCategories",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Categories/{id}/products": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "GetProductsOfCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "GetProducts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "PostProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "GetProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "operationId": "PutProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Products/{id}/picture": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "GetPictureById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Picture"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Picture"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Picture"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CustomerAddress": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "addressId": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          }
        },
        "additionalProperties": false
      },
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "countryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "stateProvinceId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "address1": {
            "type": "string",
            "nullable": true
          },
          "address2": {
            "type": "string",
            "nullable": true
          },
          "zipPostalCode": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "faxNumber": {
            "type": "string",
            "nullable": true
          },
          "customAttributes": {
            "type": "string",
            "nullable": true
          },
          "createdOnUtc": {
            "type": "string",
            "format": "date-time"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "stateProvince": {
            "$ref": "#/components/schemas/StateProvince"
          },
          "customerAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerAddress"
            },
            "nullable": true
          },
          "customerBillingAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            },
            "nullable": true
          },
          "customerShippingAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            },
            "nullable": true
          },
          "orderBillingAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          },
          "orderShippingAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "orderGuid": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "billingAddressId": {
            "type": "integer",
            "format": "int32"
          },
          "shippingAddressId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "orderTotal": {
            "type": "number",
            "format": "double"
          },
          "deleted": {
            "type": "boolean"
          },
          "createdOnUtc": {
            "type": "string",
            "format": "date-time"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "orderItemGuid": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPriceInclTax": {
            "type": "number",
            "format": "double"
          },
          "attributeDescription": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "additionalProperties": false
      },
      "Picture": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "pictureBinary": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "seoFilename": {
            "type": "string",
            "nullable": true
          },
          "altAttribute": {
            "type": "string",
            "nullable": true
          },
          "titleAttribute": {
            "type": "string",
            "nullable": true
          },
          "isNew": {
            "type": "boolean"
          },
          "productPictureMappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductPictureMapping"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductPictureMapping": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "pictureId": {
            "type": "integer",
            "format": "int32"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "picture": {
            "$ref": "#/components/schemas/Picture"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "additionalProperties": false
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "parentGroupedProductId": {
            "type": "integer",
            "format": "int32"
          },
          "visibleIndividually": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortDescription": {
            "type": "string",
            "nullable": true
          },
          "fullDescription": {
            "type": "string",
            "nullable": true
          },
          "adminComment": {
            "type": "string",
            "nullable": true
          },
          "productTemplateId": {
            "type": "integer",
            "format": "int32"
          },
          "showOnHomePage": {
            "type": "boolean"
          },
          "metaKeywords": {
            "type": "string",
            "nullable": true
          },
          "metaDescription": {
            "type": "string",
            "nullable": true
          },
          "metaTitle": {
            "type": "string",
            "nullable": true
          },
          "isRental": {
            "type": "boolean"
          },
          "rentalPriceLength": {
            "type": "integer",
            "format": "int32"
          },
          "rentalPricePeriodId": {
            "type": "integer",
            "format": "int32"
          },
          "isTaxExempt": {
            "type": "boolean"
          },
          "taxCategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "displayStockAvailability": {
            "type": "boolean"
          },
          "displayStockQuantity": {
            "type": "boolean"
          },
          "minStockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "orderMinimumQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "orderMaximumQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "disableBuyButton": {
            "type": "boolean"
          },
          "disableWishlistButton": {
            "type": "boolean"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "oldPrice": {
            "type": "number",
            "format": "double"
          },
          "specialPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "specialPriceStartDateTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "specialPriceEndDateTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "length": {
            "type": "number",
            "format": "double"
          },
          "width": {
            "type": "number",
            "format": "double"
          },
          "height": {
            "type": "number",
            "format": "double"
          },
          "availableStartDateTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "availableEndDateTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "published": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "createdOnUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedOnUtc": {
            "type": "string",
            "format": "date-time"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "nullable": true
          },
          "productCategoryMappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCategoryMapping"
            },
            "nullable": true
          },
          "productPictureMappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductPictureMapping"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductCategoryMapping": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "isFeaturedProduct": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "additionalProperties": false
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "categoryTemplateId": {
            "type": "integer",
            "format": "int32"
          },
          "metaKeywords": {
            "type": "string",
            "nullable": true
          },
          "metaDescription": {
            "type": "string",
            "nullable": true
          },
          "metaTitle": {
            "type": "string",
            "nullable": true
          },
          "parentCategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "pictureId": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "allowCustomersToSelectPageSize": {
            "type": "boolean"
          },
          "pageSizeOptions": {
            "type": "string",
            "nullable": true
          },
          "priceRanges": {
            "type": "string",
            "nullable": true
          },
          "showOnHomePage": {
            "type": "boolean"
          },
          "includeInTopMenu": {
            "type": "boolean"
          },
          "hasDiscountsApplied": {
            "type": "boolean"
          },
          "subjectToAcl": {
            "type": "boolean"
          },
          "limitedToStores": {
            "type": "boolean"
          },
          "published": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "createdOnUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedOnUtc": {
            "type": "string",
            "format": "date-time"
          },
          "productCategoryMappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCategoryMapping"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}